+++ src/validators/FieldValidator.php
@@ -158,6 +158,12 @@ class FieldValidator extends Validator
                 );
             }
 
+            if (!isset($blockSiblingCount[$parentId][$block->typeId])) {
+                // This block is in a place in the field where it isn't supposed to exist any more, so ignore it for the
+                // purposes of min sibling blocks validation
+                continue;
+            }
+
             $blockSiblingCount[$parentId][$block->typeId] += 1;
             $lastBlock = $block;
         }