+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased (4.x)
+
+### Changed
+- Whether a Neo block type's top level setting has been violated is now validated server-side when saving a Neo field's contents, rather than relying on it to be enforced by client-side JavaScript
+
 ## Unreleased
 
 ### Fixed